reafactor : PrimaryViewController에서 SecondaryViewController로 데이터 전달하는 로직 수정
- PrimaryViewController 의 giveDataToSecondaryVC 메소드 수정
- SecondaryViewController에 delegate 채택
- TableViewIdexPathHolder에 textView의 text 속성 초기화할 속성 추가
- TextViewDelegate 파일 삭제
